okh: Removes JSON-Schema
because it does not describe OKH v1,
but rather Helpful engineering's extension of it.
In a meeting in May,
we (including Helpful Engineering, IoPA, OSEG and others)
decided to switch to OKH-LOSH,
and then include Helpfuls additions in there.

The conforming OKH v1 JSON-Schema is not in the OpenKnowHow repo:
https://github.com/iop-alliance/OpenKnowHow/blob/master/src/schema/okh-v1.schema.json